backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
langchain-anthropic==0.3.12
langchain-core==0.3.51
langgraph==0.4.1
